Improved logging for CVFM sync
Partial-Jira-Bug: CEM-6537

+++ cvfm/synchronizers.py
@@ -1,3 +1,8 @@
+import logging
+
+logger = logging.getLogger(__name__)
+
+
 class Synchronizer(object):
     def __init__(
         self,
@@ -29,7 +34,9 @@ class VirtualMachineSynchronizer(object):
         self._vm_service = vm_service
 
     def sync(self):
+        logger.info("Populating local database with VM models...")
         self._vm_service.populate_db_with_vms()
+        logger.info("Populated local database with VM models")
 
 
 class DistributedPortGroupSynchronizer(object):
@@ -37,7 +44,9 @@ class DistributedPortGroupSynchronizer(object):
         self._dpg_service = dpg_service
 
     def sync_create(self):
+        logger.info("Populating local database with DPG models...")
         self._dpg_service.populate_db_with_dpgs()
+        logger.info("Populated local database with DPG models")
         dpgs_in_vcenter = self._dpg_service.get_all_dpg_models()
         fabric_vns = self._dpg_service.get_all_fabric_vns()
         fabric_vn_uuids = [vn.uuid for vn in fabric_vns]
@@ -48,8 +57,14 @@ class DistributedPortGroupSynchronizer(object):
             if dpg_model.uuid not in fabric_vn_uuids
         ]
 
+        if len(vns_to_create) == 0:
+            logger.info("Not detected lacking VNs in VNC")
+            return
+
+        logger.info("Creating lacking VNs in VNC...")
         for dpg_model in vns_to_create:
             self._dpg_service.create_fabric_vn(dpg_model)
+        logger.info("Created lacking VNs in VNC")
 
     def sync_delete(self):
         dpgs_in_vcenter = self._dpg_service.get_all_dpg_models()
@@ -62,8 +77,14 @@ class DistributedPortGroupSynchronizer(object):
             if fabric_vn.uuid not in dpgs_in_vcenter_uuids
         ]
 
+        if len(fabric_vns_to_delete) == 0:
+            logger.info("Not detected stale VNs in VNC")
+            return
+
+        logger.info("Deleting stale VNs from VNC...")
         for fabric_vn in fabric_vns_to_delete:
             self._dpg_service.delete_fabric_vn(fabric_vn.uuid)
+        logger.info("Deleted stale VNs from VNC...")
 
 
 class VirtualPortGroupSynchronizer(object):
@@ -72,13 +93,17 @@ class VirtualPortGroupSynchronizer(object):
         self._vpg_service = vpg_service
 
     def sync_create(self):
+        logger.info("Creating lacking/Updating VPGs in VNC...")
         vm_models = self._vm_service.get_all_vm_models()
         vpg_models = []
         for vm_model in vm_models:
             vpg_models.extend(self._vpg_service.create_vpg_models(vm_model))
+        logger.debug("Creating/Updating VPGs in VNC ")
         for vpg_model in set(vpg_models):
+            logger.debug("Syncing VPG in VNC for %s", vpg_model)
             self._vpg_service.create_vpg_in_vnc(vpg_model)
             self._vpg_service.attach_pis_to_vpg(vpg_model)
+        logger.info("Created lacking/Updated VPGs in VNC")
 
     def sync_delete(self):
         vm_models = self._vm_service.get_all_vm_models()
@@ -90,8 +115,13 @@ class VirtualPortGroupSynchronizer(object):
             vpg.uuid for vpg in self._vpg_service.read_all_vpgs()
         )
         vpgs_to_delete = vpg_uuids_in_vnc - vpg_uuids
+        if len(vpgs_to_delete) == 0:
+            logger.info("Not detected stale VPGs in VNC")
+            return
+        logger.info("Deleting stale VPGs from VNC...")
         for vpg_uuid in vpgs_to_delete:
             self._vpg_service.delete_vpg(vpg_uuid)
+        logger.info("Deleted stale VPGs from VNC...")
 
 
 class VirtualMachineInterfaceSynchronizer(object):
@@ -106,9 +136,12 @@ class VirtualMachineInterfaceSynchronizer(object):
             vmi_models.extend(
                 self._vmi_service.create_vmi_models_for_vm(vm_model)
             )
+        logger.info("Creating lacking/Updating VMIs in VNC...")
         for vmi_model in set(vmi_models):
+            logger.debug("Syncing VMI in VNC for %s", vmi_model)
             self._vmi_service.create_vmi_in_vnc(vmi_model)
             self._vmi_service.attach_vmi_to_vpg(vmi_model)
+        logger.info("Created lacking/Updated VMIs in VNC")
 
     def sync_delete(self):
         vm_models = self._vm_service.get_all_vm_models()
@@ -122,5 +155,11 @@ class VirtualMachineInterfaceSynchronizer(object):
             vmi.uuid for vmi in self._vmi_service.read_all_vmis()
         )
         vmis_to_delete = vmi_uuids_in_vnc - vmi_uuids
+
+        if len(vmis_to_delete) == 0:
+            logger.info("Not detected stale VMIs in VNC")
+            return
+        logger.info("Deleting stale VMIs from VNC...")
         for vmi_uuid in vmis_to_delete:
             self._vmi_service.delete_vmi(vmi_uuid)
+        logger.info("Deleted stale VMIs from VNC")
